requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
django-phone-field==1.8.1
pytz==2021.1
sqlparse==0.4.1
